refactor(Authentication): organize module and adopt Swift API design guidelines
