+++ Blockchain/Authentication/Models/Passcode.swift
@@ -1,0 +1,14 @@
+//
+//  Passcode.swift
+//  Blockchain
+//
+//  Created by Maurice A. on 4/13/18.
+//  Copyright © 2018 Blockchain Luxembourg S.A. All rights reserved.
+//
+
+import Foundation
+
+/// Represents a passcode used for authenticating the user.
+internal struct Passcode {
+    // TODO: implement passcode structure
+}
